+++ nproto/natsrpc/natsrpc_test.go
@@ -1,4 +1,4 @@
-package nprpc
+package natsrpc
 
 import (
 	"context"
