+++ tests/runner/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/golang/protobuf/ptypes/duration"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/huangjunwen/nproto/helpers/taskrunner"
-	"github.com/huangjunwen/nproto/nproto/nprpc"
+	"github.com/huangjunwen/nproto/nproto/natsrpc"
 	nats "github.com/nats-io/nats.go"
 
 	runnerapi "github.com/huangjunwen/nproto/tests/runner/api"
@@ -44,9 +44,9 @@ func main() {
 	defer runner.Close()
 
 	// Creates the server.
-	server, err := nprpc.NewNatsRPCServer(
+	server, err := natsrpc.NewNatsRPCServer(
 		nc,
-		nprpc.ServerOptTaskRunner(runner),
+		natsrpc.ServerOptTaskRunner(runner),
 	)
 	if err != nil {
 		log.Panic(err)
@@ -61,7 +61,7 @@ func main() {
 	log.Printf("Svc registered.\n")
 
 	// Creates the client.
-	client, err := nprpc.NewNatsRPCClient(nc)
+	client, err := natsrpc.NewNatsRPCClient(nc)
 	if err != nil {
 		log.Panic(err)
 	}
